+++ packages/brick-kit/src/core/Router.ts
@@ -39,6 +39,7 @@ import { preCheckPermissions } from "../internal/checkPermissions";
 import { clearPollTimeout } from "../internal/poll";
 import { shouldBeDefaultCollapsed } from "../internal/shouldBeDefaultCollapsed";
 import { CustomTemplateContext } from "./CustomTemplates";
+import { registerStoryboardFunctions } from "./StoryboardFunctions";
 
 export class Router {
   private defaultCollapsed = false;
@@ -216,8 +217,9 @@ export class Router {
         this.locationContext.getSubStoryboardByRoute(storyboard);
       await this.kernel.loadDepsOfStoryboard(subStoryboard);
 
-      // 注册 Storyboard 中定义的自定义模板。
+      // 注册 Storyboard 中定义的自定义模板和函数。
       this.kernel.registerCustomTemplatesInStoryboard(storyboard);
+      registerStoryboardFunctions(storyboard);
     }
 
     const { mountPoints, currentApp: previousApp } = this.kernel;