+++ packages/brick-scripts/src/ask.spec.ts
@@ -4,7 +4,6 @@ import { TargetType } from "./interface";
 import { askTargetType } from "./questions/askTargetType";
 import { askPackageName } from "./questions/askPackageName";
 import { askBrickName } from "./questions/askBrickName";
-import { askTemplateName } from "./questions/askTemplateName";
 import { askProcessorName } from "./questions/askProcessorName";
 import { askEditorBrickName } from "./questions/askEditorBrickName";
 import { ask } from "./ask";
@@ -14,7 +13,6 @@ jest.mock("@next-core/repo-config");
 jest.mock("./questions/askTargetType");
 jest.mock("./questions/askPackageName");
 jest.mock("./questions/askBrickName");
-jest.mock("./questions/askTemplateName");
 jest.mock("./questions/askProcessorName");
 jest.mock("./questions/askEditorBrickName");
 
@@ -25,7 +23,6 @@ jest.mock("./questions/askEditorBrickName");
 const spyOnAskTargetType = askTargetType as jest.Mock;
 const spyOnAskPackageName = askPackageName as jest.Mock;
 const spyOnAskBrickName = askBrickName as jest.Mock;
-const spyOnAskTemplateName = askTemplateName as jest.Mock;
 const spyOnAskProcessorName = askProcessorName as jest.Mock;
 const spyOnAskEditorBrickName = askEditorBrickName as jest.Mock;
 
@@ -39,34 +36,24 @@ describe("ask", () => {
     jest.resetAllMocks();
   });
 
-  test.each<[TargetType, string, string, string, string, string]>([
-    [TargetType.A_NEW_BRICK, "package-a", "brick-a", "", "", ""],
-    [TargetType.A_NEW_EDITOR_BRICK, "package-a", "", "", "", "brick-a"],
-    [
-      TargetType.A_NEW_CUSTOM_PROVIDER_BRICK,
-      "package-a",
-      "brick-b",
-      "",
-      "",
-      "",
-    ],
-    [TargetType.A_NEW_CUSTOM_PROCESSOR, "package-a", "", "", "doSomething", ""],
-    [TargetType.A_NEW_PACKAGE_OF_LIBS, "lib-a", "", "", "", ""],
-    [TargetType.A_NEW_LEGACY_TEMPLATE, "package-b", "", "template-b", "", ""],
+  test.each<[TargetType, string, string, string, string]>([
+    [TargetType.A_NEW_BRICK, "package-a", "brick-a", "", ""],
+    [TargetType.A_NEW_EDITOR_BRICK, "package-a", "", "", "brick-a"],
+    [TargetType.A_NEW_CUSTOM_PROVIDER, "package-a", "brick-b", "", ""],
+    [TargetType.A_NEW_CUSTOM_PROCESSOR, "package-a", "", "doSomething", ""],
+    [TargetType.A_NEW_PACKAGE_OF_LIBS, "lib-a", "", "", ""],
   ])(
-    "should return { targetType: '%s', packageName: '%s', brickName: '%s', templateName: '%s', processorName: '%s' }",
+    "should return { targetType: '%s', packageName: '%s', brickName: '%s', processorName: '%s' }",
     async (
       targetType,
       packageName,
       brickName,
-      templateName,
       processorName,
       editorBrickName
     ) => {
       spyOnAskTargetType.mockReturnValue({ targetType });
       spyOnAskPackageName.mockReturnValue({ packageName });
       spyOnAskBrickName.mockReturnValue({ brickName });
-      spyOnAskTemplateName.mockReturnValue({ templateName });
       spyOnAskProcessorName.mockReturnValue({ processorName });
       spyOnAskEditorBrickName.mockReturnValue({ brickName: editorBrickName });
 
@@ -74,7 +61,6 @@ describe("ask", () => {
         targetType,
         packageName,
         brickName: editorBrickName || brickName,
-        templateName,
         processorName,
       });
     }
@@ -83,7 +69,9 @@ describe("ask", () => {
   it("should use specified type", async () => {
     spyOnAskTargetType.mockReturnValue({ choices: [{ value: "libs" }] });
 
-    expect(await ask("", { type: "libs" })).toMatchObject({
+    expect(
+      await ask("", { type: TargetType.A_NEW_PACKAGE_OF_LIBS })
+    ).toMatchObject({
       targetType: "libs",
     });
   });