+++ packages/brick-scripts/src/questions/askPackageName.ts
@@ -22,7 +22,7 @@ function legacyAskPackageName({
       TargetType.A_NEW_BRICK,
       TargetType.A_NEW_EDITOR_BRICK,
       TargetType.A_NEW_CUSTOM_TEMPLATE,
-      TargetType.A_NEW_CUSTOM_PROVIDER_BRICK,
+      TargetType.A_NEW_CUSTOM_PROVIDER,
       TargetType.A_NEW_CUSTOM_PROCESSOR,
     ].includes(targetType)
   ) {
@@ -48,23 +48,6 @@ function legacyAskPackageName({
     };
   }
 
-  if (targetType === TargetType.A_NEW_LEGACY_TEMPLATE) {
-    // 读取当前的 `templates/*` 作为候选列表。
-    const root = path.join(appRoot, "templates");
-    const pkgList = fs
-      .readdirSync(root, { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .map((dirent) => dirent.name);
-
-    return {
-      type: "list",
-      name: "packageName",
-      message: "which package do you want to put the new template in?",
-      choices: pkgList,
-      // default: loadHistory().lastSelectedTemplatePackage,
-    };
-  }
-
   if (targetType === TargetType.A_NEW_PACKAGE_OF_PROVIDERS) {
     // 读取所有的 `sdk/*` 作为候选列表。
     const root = path.join(
@@ -84,44 +67,6 @@ function legacyAskPackageName({
       name: "packageName",
       message: "which sdk do you want to create providers for?",
       choices: sdkList,
-    };
-  }
-
-  if (targetType === TargetType.TRANSFORM_A_MICRO_APP) {
-    // 读取所有的 `micro-apps/*` 作为候选列表。
-    const root = path.join(appRoot, "micro-apps");
-    const microApps = fs
-      .readdirSync(root, { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .filter(
-        (dirent) => !fs.existsSync(path.join(root, dirent.name, "src/index.ts"))
-      )
-      .map((dirent) => dirent.name);
-
-    return {
-      type: "list",
-      name: "packageName",
-      message: "which micro-app do you want to transform?",
-      choices: microApps,
-    };
-  }
-
-  if (targetType === TargetType.I18N_PATCH_A_PACKAGE_OF_LEGACY_TEMPLATES) {
-    // 读取所有的 `templates/*` 作为候选列表。
-    const root = path.join(appRoot, "templates");
-    const microApps = fs
-      .readdirSync(root, { withFileTypes: true })
-      .filter((dirent) => dirent.isDirectory())
-      .filter(
-        (dirent) => !fs.existsSync(path.join(root, dirent.name, "src/i18n"))
-      )
-      .map((dirent) => dirent.name);
-
-    return {
-      type: "list",
-      name: "packageName",
-      message: "which package do you want to patch?",
-      choices: microApps,
     };
   }