+++ packages/brick-types/src/index.ts
@@ -10,3 +10,4 @@ export * from "./menu";
 export * from "./runtime";
 export * from "./story";
 export * from "./builder";
+export * from "./utility";