+++ packages/brick-kit/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.66.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.65.2...@next-core/brick-kit@2.66.0) (2021-08-30)
+
+
+### Features
+
+* support storyboard functions ([b27c235](https://github.com/easyops-cn/next-core/commit/b27c2354eb99fa82e515526e605c368f0dcca20a))
+
+
+
+
+
+## [2.65.2](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.65.1...@next-core/brick-kit@2.65.2) (2021-08-27)
+
+**Note:** Version bump only for package @next-core/brick-kit
+
+
+
+
+
+## [2.65.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.65.0...@next-core/brick-kit@2.65.1) (2021-08-24)
+
+
+### Bug Fixes
+
+* fix dynamic flow api initializing ([0f3c2fe](https://github.com/easyops-cn/next-core/commit/0f3c2fe01e9b15707491f503780c9c161107ea86))
+
+
+
+
+
 # [2.65.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-kit@2.64.1...@next-core/brick-kit@2.65.0) (2021-08-17)