+++ packages/dev-dependencies/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.24](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.23...@next-core/dev-dependencies@1.10.24) (2021-08-30)
+
+**Note:** Version bump only for package @next-core/dev-dependencies
+
+
+
+
+
+## [1.10.23](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.22...@next-core/dev-dependencies@1.10.23) (2021-08-30)
+
+**Note:** Version bump only for package @next-core/dev-dependencies
+
+
+
+
+
+## [1.10.22](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.21...@next-core/dev-dependencies@1.10.22) (2021-08-27)
+
+**Note:** Version bump only for package @next-core/dev-dependencies
+
+
+
+
+
+## [1.10.21](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.20...@next-core/dev-dependencies@1.10.21) (2021-08-24)
+
+**Note:** Version bump only for package @next-core/dev-dependencies
+
+
+
+
+
+## [1.10.20](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.19...@next-core/dev-dependencies@1.10.20) (2021-08-18)
+
+**Note:** Version bump only for package @next-core/dev-dependencies
+
+
+
+
+
+## [1.10.19](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.18...@next-core/dev-dependencies@1.10.19) (2021-08-18)
+
+**Note:** Version bump only for package @next-core/dev-dependencies
+
+
+
+
+
 ## [1.10.18](https://github.com/easyops-cn/next-core/compare/@next-core/dev-dependencies@1.10.17...@next-core/dev-dependencies@1.10.18) (2021-08-17)
 
 **Note:** Version bump only for package @next-core/dev-dependencies